Expose casting blocks in header
